+++ release/ray_release/aws.py
@@ -1,7 +1,9 @@
 import os
 from copy import deepcopy
+from typing import Optional
 
 import boto3
+from botocore.exceptions import ClientError
 from ray_release.logger import logger
 from ray_release.util import DeferredEnvVar
 
@@ -72,3 +74,30 @@ def add_tags_to_aws_config(aws_config: dict, tags_to_add: dict, resource_types: 
             tags.append({"Key": key, "Value": value})
 
     return aws_config
+
+
+def upload_to_s3(src_path: str, bucket: str, key_path: str) -> Optional[str]:
+    """Upload a file to a S3 bucket
+
+    This assumes the bucket has public read access on the objects uploaded.
+
+    Args:
+        src_path: local file path.
+        bucket: S3 bucket name.
+        key_path: destination url of the uploaded object.
+
+    Return:
+        HTTP URL where the uploaded object could be downloaded if successful,
+        or None if fails.
+
+    Raises:
+        ClientError if upload fails
+    """
+    s3_client = boto3.client("s3")
+    try:
+        s3_client.upload_file(Filename=src_path, Bucket=bucket, Key=key_path)
+    except ClientError as e:
+        logger.warning(f"Failed to upload to s3:  {e}")
+        return None
+
+    return f"https://{bucket}.s3.us-west-2.amazonaws.com/{key_path}"
