+++ release/ray_release/wheels.py
@@ -4,11 +4,10 @@ import re
 import shlex
 import subprocess
 import sys
+from datetime import datetime
 import tempfile
 import time
 import urllib.request
-import boto3
-from botocore.exceptions import ClientError
 from typing import Optional, List, Tuple
 
 from ray_release.config import DEFAULT_PYTHON_VERSION, parse_python_version
@@ -21,6 +20,7 @@ from ray_release.exception import (
 )
 from ray_release.logger import logger
 from ray_release.util import url_exists, python_version_str, resolve_url
+from ray_release.aws import upload_to_s3
 
 DEFAULT_BRANCH = "master"
 DEFAULT_GIT_OWNER = "ray-project"
@@ -133,22 +133,19 @@ def parse_wheels_filename(
     return ray_version, python_version
 
 
-def upload_to_s3(src_path: str, bucket: str, key_path: str) -> Optional[str]:
-    s3_client = boto3.client("s3")
-    try:
-        s3_client.upload_file(Filename=src_path, Bucket=bucket, Key=key_path)
-    except ClientError as e:
-        logger.warning(f"Failed to upload to s3:  {e}")
-        return None
-
-    return f"https://{bucket}.s3.us-west-2.amazonaws.com/{key_path}"
-
-
 def get_ray_wheels_url_from_local_wheel(ray_wheels: str) -> Optional[str]:
-    """Upload a local wheel file to S3 and return the downloadable URI"""
-    import os
-    from datetime import datetime
+    """Upload a local wheel file to S3 and return the downloadable URI
 
+    The uploaded object will have local user and current timestamp encoded
+    in the upload key path, e.g.:
+    "ubuntu_2022_01_01_23:59:99/ray-3.0.0.dev0-cp37-cp37m-manylinux_x86_64.whl"
+
+    Args:
+        ray_wheels: File path with `file://` prefix.
+
+    Return:
+        Downloadable HTTP URL to the uploaded wheel on S3.
+    """
     wheel_path = ray_wheels[len("file://") :]
     wheel_name = os.path.basename(wheel_path)
     if not os.path.exists(wheel_path):
@@ -157,7 +154,7 @@ def get_ray_wheels_url_from_local_wheel(ray_wheels: str) -> Optional[str]:
 
     bucket = RELEASE_MANUAL_WHEEL_BUCKET
     unique_dest_path_prefix = (
-        f'{os.getlogin()}_{datetime.now().strftime("%Y-%m-%d-%H%MZ")}'
+        f'{os.getlogin()}_{datetime.now().strftime("%Y_%m_%d_%H:%M:%S")}'
     )
     key_path = f"{unique_dest_path_prefix}/{wheel_name}"
     return upload_to_s3(wheel_path, bucket, key_path)
